Factor transport out into its own module
